+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## 1.5.0 (May 7, 2013)
+
+Features:
+
+  - Added "prev" and "next" links support
+  - Added refresh meta tag support
+
+## 1.4.1 (March 14, 2013)
+
+Bugfixes:
+
+  - Added support for Hash inside of an Array
+
+## 1.4.0 (March 14, 2013)
+
+Features:
+
+  - Added support of custom meta tags
+
+## 1.3.0 (February 13, 2013)
+
+Features:
+
+  - Added Hash and Array as possible contents for the meta tags. Check README for details
+  - Added support of string meta tag names
+  - Allow to disable noindex and nofollow using `false` as a value
+
+Bugfixes:
+
+  - Do not display title HTML tag when title is blank
+  - Do not display OpenGraph tags when content is empty
+
 ## 1.2.6 (March 4, 2012)
 
 Features: